search_engine/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
torch==1.9.0
transformers==4.9.2
qdrant-client==0.7.0
Pillow==8.3.1
python-multipart==0.0.5
pydantic==1.8.2
pandas==1.3.5